% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rules_to_case_when}
\alias{rules_to_case_when}
\title{rules_to_case_when}
\usage{
rules_to_case_when(corels_output)
}
\arguments{
\item{corels_output}{corels captured output}
}
\description{
converts captured corels output to dplyr::Case_when()
}
\examples{
# not run:
}
\keyword{capture}
\keyword{corels}
