% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{write_to_txt_file}
\alias{write_to_txt_file}
\title{write_to_txt_file}
\usage{
write_to_txt_file(df, cols, txt_file)
}
\arguments{
\item{df}{name of the dataframe to convert}

\item{cols}{a list of the columns in the data frame to convert}

\item{txt_file}{the name of the text file}
}
\description{
converts columns of a dataframe to a textfile in the format corels can use
}
\examples{
# not run:
# write_to_txt_file(df = df, cols = cols, txt_file = "train.txt")
}
\keyword{convert}
\keyword{dataframe}
\keyword{textfile}
