% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{predict_corels}
\alias{predict_corels}
\title{predict_corels}
\usage{
predict_corels(model, new_df)
}
\arguments{
\item{model}{R object created by tidy_corels() function.}

\item{new_df}{A new dataframe to apply corels rules to and generate a classification.}
}
\description{
Applies the Corels rules in theR object returned by tidycorels::corels() to a new dataframe. For example, test data not used to create the Corels rules.
}
\examples{
# See <https://github.com/billster45/tidycorels/blob/master/README.md>
}
